MarketPulsePro/requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
aiohttp==3.9.1
aiosqlite==0.19.0
sqlalchemy==2.0.23
apscheduler==3.10.4
requests==2.31.0
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
plotly==5.18.0
cachetools==5.3.2
pytz==2023.3
beautifulsoup4==4.12.2
feedparser==6.0.10
pydantic==2.5.0
redis==5.0.1
celery==5.3.4
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0